Colon and Rectal Cancer Clinical and Virome Data

A dataset containing cleaned key clinical, survival, and virome abundance data for patients with colon and rectal adenocarcinoma from the TCGA PanCancer Atlas study.

Usage

crc_virome

Format

A data frame with patient data in rows and variables in columns:

sample_id

Unique TCGA patient identifier.

time_months

Overall survival time in months, censored at 60 months.

status

Overall survival status: 1 for deceased, 0 for alive.

age

Age of the patient at diagnosis.

sex

Sex of the patient.

race

Race of the patient.

ethnicity

Ethnicity of the patient.

stage

AJCC pathologic tumor stage.

Alfamovirus

Normalised abundance for the Alfamovirus genus.

Allexivirus

Normalised abundance for the Allexivirus genus.

Alphacoronavirus

Normalised abundance for the Alphacoronavirus genus.

Alphaentomopoxvirus

Normalised abundance for the Alphaentomopoxvirus genus.

Alphafusellovirus

Normalised abundance for the Alphafusellovirus genus.

Alphapapillomavirus

Normalised abundance for the Alphapapillomavirus genus.

Alphapartitivirus

Normalised abundance for the Alphapartitivirus genus.

Alpharetrovirus

Normalised abundance for the Alpharetrovirus genus.

Alphatorquevirus

Normalised abundance for the Alphatorquevirus genus.

Alphavirus

Normalised abundance for the Alphavirus genus.

Ambidensovirus

Normalised abundance for the Ambidensovirus genus.

Ampelovirus

Normalised abundance for the Ampelovirus genus.

Andromedalikevirus

Normalised abundance for the Andromedalikevirus genus.

Aparavirus

Normalised abundance for the Aparavirus genus.

Aquamavirus

Normalised abundance for the Aquamavirus genus.

Arenavirus

Normalised abundance for the Arenavirus genus.

Avastrovirus

Normalised abundance for the Avastrovirus genus.

Bacilladnavirus

Normalised abundance for the Bacilladnavirus genus.

Bafinivirus

Normalised abundance for the Bafinivirus genus.

Barnyardlikevirus

Normalised abundance for the Barnyardlikevirus genus.

Batrachovirus

Normalised abundance for the Batrachovirus genus.

Bcep22likevirus

Normalised abundance for the Bcep22likevirus genus.

Betacoronavirus

Normalised abundance for the Betacoronavirus genus.

Betaentomopoxvirus

Normalised abundance for the Betaentomopoxvirus genus.

Betapapillomavirus

Normalised abundance for the Betapapillomavirus genus.

Betapartitivirus

Normalised abundance for the Betapartitivirus genus.

Betaretrovirus

Normalised abundance for the Betaretrovirus genus.

Betatorquevirus

Normalised abundance for the Betatorquevirus genus.

Bicaudavirus

Normalised abundance for the Bicaudavirus genus.

Bracovirus

Normalised abundance for the Bracovirus genus.

Bromovirus

Normalised abundance for the Bromovirus genus.

Cafeteriavirus

Normalised abundance for the Cafeteriavirus genus.

Capripoxvirus

Normalised abundance for the Capripoxvirus genus.

Carlavirus

Normalised abundance for the Carlavirus genus.

Caulimovirus

Normalised abundance for the Caulimovirus genus.

Cavemovirus

Normalised abundance for the Cavemovirus genus.

Cervidpoxvirus

Normalised abundance for the Cervidpoxvirus genus.

Chlorovirus

Normalised abundance for the Chlorovirus genus.

Closterovirus

Normalised abundance for the Closterovirus genus.

Comovirus

Normalised abundance for the Comovirus genus.

Cp220likevirus

Normalised abundance for the Cp220likevirus genus.

Crinivirus

Normalised abundance for the Crinivirus genus.

Cripavirus

Normalised abundance for the Cripavirus genus.

Cucumovirus

Normalised abundance for the Cucumovirus genus.

Cytomegalovirus

Normalised abundance for the Cytomegalovirus genus.

Deltabaculovirus

Normalised abundance for the Deltabaculovirus genus.

Dicipivirus

Normalised abundance for the Dicipivirus genus.

Dyopipapillomavirus

Normalised abundance for the Dyopipapillomavirus genus.

Emaravirus

Normalised abundance for the Emaravirus genus.

Endornavirus

Normalised abundance for the Endornavirus genus.

Enterovirus

Normalised abundance for the Enterovirus genus.

Fabavirus

Normalised abundance for the Fabavirus genus.

Flavivirus

Normalised abundance for the Flavivirus genus.

Furovirus

Normalised abundance for the Furovirus genus.

Gammacoronavirus

Normalised abundance for the Gammacoronavirus genus.

Gammapapillomavirus

Normalised abundance for the Gammapapillomavirus genus.

Gammaretrovirus

Normalised abundance for the Gammaretrovirus genus.

Gammatorquevirus

Normalised abundance for the Gammatorquevirus genus.

Hepacivirus

Normalised abundance for the Hepacivirus genus.

Hepatovirus

Normalised abundance for the Hepatovirus genus.

Higrevirus

Normalised abundance for the Higrevirus genus.

Hordeivirus

Normalised abundance for the Hordeivirus genus.

Hypovirus

Normalised abundance for the Hypovirus genus.

I3likevirus

Normalised abundance for the I3likevirus genus.

Ichnovirus

Normalised abundance for the Ichnovirus genus.

Iflavirus

Normalised abundance for the Iflavirus genus.

Ilarvirus

Normalised abundance for the Ilarvirus genus.

Influenzavirus_C

Normalised abundance for the Influenzavirus_C genus.

Isavirus

Normalised abundance for the Isavirus genus.

Kobuvirus

Normalised abundance for the Kobuvirus genus.

L5likevirus

Normalised abundance for the L5likevirus genus.

Lambdalikevirus

Normalised abundance for the Lambdalikevirus genus.

Lambdapapillomavirus

Normalised abundance for the Lambdapapillomavirus genus.

Lentivirus

Normalised abundance for the Lentivirus genus.

Leporipoxvirus

Normalised abundance for the Leporipoxvirus genus.

Lymphocryptovirus

Normalised abundance for the Lymphocryptovirus genus.

Mamastrovirus

Normalised abundance for the Mamastrovirus genus.

Mardivirus

Normalised abundance for the Mardivirus genus.

Mastadenovirus

Normalised abundance for the Mastadenovirus genus.

Microvirus

Normalised abundance for the Microvirus genus.

Mimivirus

Normalised abundance for the Mimivirus genus.

Molluscipoxvirus

Normalised abundance for the Molluscipoxvirus genus.

Muromegalovirus

Normalised abundance for the Muromegalovirus genus.

Muscavirus

Normalised abundance for the Muscavirus genus.

N4likevirus

Normalised abundance for the N4likevirus genus.

Narnavirus

Normalised abundance for the Narnavirus genus.

Nepovirus

Normalised abundance for the Nepovirus genus.

Nyavirus

Normalised abundance for the Nyavirus genus.

Omegapapillomavirus

Normalised abundance for the Omegapapillomavirus genus.

Orthobunyavirus

Normalised abundance for the Orthobunyavirus genus.

Orthohepadnavirus

Normalised abundance for the Orthohepadnavirus genus.

Orthopoxvirus

Normalised abundance for the Orthopoxvirus genus.

Ostreavirus

Normalised abundance for the Ostreavirus genus.

Parapoxvirus

Normalised abundance for the Parapoxvirus genus.

Pecluvirus

Normalised abundance for the Pecluvirus genus.

Pestivirus

Normalised abundance for the Pestivirus genus.

Phi29likevirus

Normalised abundance for the Phi29likevirus genus.

Phikmvlikevirus

Normalised abundance for the Phikmvlikevirus genus.

Phikzlikevirus

Normalised abundance for the Phikzlikevirus genus.

Piscihepevirus

Normalised abundance for the Piscihepevirus genus.

Pithovirus

Normalised abundance for the Pithovirus genus.

Polyomavirus

Normalised abundance for the Polyomavirus genus.

Pomovirus

Normalised abundance for the Pomovirus genus.

Potexvirus

Normalised abundance for the Potexvirus genus.

Potyvirus

Normalised abundance for the Potyvirus genus.

Prasinovirus

Normalised abundance for the Prasinovirus genus.

Proboscivirus

Normalised abundance for the Proboscivirus genus.

Protoparvovirus

Normalised abundance for the Protoparvovirus genus.

Prymnesiovirus

Normalised abundance for the Prymnesiovirus genus.

Ranavirus

Normalised abundance for the Ranavirus genus.

Rhadinovirus

Normalised abundance for the Rhadinovirus genus.

Rubulavirus

Normalised abundance for the Rubulavirus genus.

Salivirus

Normalised abundance for the Salivirus genus.

Sapelovirus

Normalised abundance for the Sapelovirus genus.

Sclerodarnavirus

Normalised abundance for the Sclerodarnavirus genus.

Senecavirus

Normalised abundance for the Senecavirus genus.

Sfi1unalikevirus

Normalised abundance for the Sfi1unalikevirus genus.

Sfi21dtunalikevirus

Normalised abundance for the Sfi21dtunalikevirus genus.

Sicinivirus

Normalised abundance for the Sicinivirus genus.

Simplexvirus

Normalised abundance for the Simplexvirus genus.

Skunalikevirus

Normalised abundance for the Skunalikevirus genus.

Sp6likevirus

Normalised abundance for the Sp6likevirus genus.

Spo1virus

Normalised abundance for the Spo1virus genus.

Spounalikevirus

Normalised abundance for the Spounalikevirus genus.

T4likevirus

Normalised abundance for the T4likevirus genus.

T5likevirus

Normalised abundance for the T5likevirus genus.

Taupapillomavirus

Normalised abundance for the Taupapillomavirus genus.

Tenuivirus

Normalised abundance for the Tenuivirus genus.

Tobamovirus

Normalised abundance for the Tobamovirus genus.

Totivirus

Normalised abundance for the Totivirus genus.

Trichovirus

Normalised abundance for the Trichovirus genus.

Tritimovirus

Normalised abundance for the Tritimovirus genus.

Tunalikevirus

Normalised abundance for the Tunalikevirus genus.

Tymovirus

Normalised abundance for the Tymovirus genus.

Vesivirus

Normalised abundance for the Vesivirus genus.

Waikavirus

Normalised abundance for the Waikavirus genus.

Whispovirus

Normalised abundance for the Whispovirus genus.

Yatapoxvirus

Normalised abundance for the Yatapoxvirus genus.

Source

The data was sourced, cleaned, and merged from the supplementary materials of Smyth et al. (2024) [https://doi.org/10.1002/cam4.70434], originating from the cBioPortal for Cancer Genomics ('TCGA, PanCancer Atlas' study).