Simulated Rapeseed Germination and Growth Data

Simulated germination and early growth of rapeseed (Brassica napus L.) across a range of temperatures. Based on parameters from Haj Sghaier et al. (2022).

Format

A data frame with 1040 rows and 5 variables:

temperature

Experimental temperature (Celsius) (numeric).

replicate

Experimental replicate number (integer).

time

Day of measurement (integer).

growth

Simulated seedling growth (numeric).

germinated

Germination status: 1 = germinated, 0 = not germinated (integer).

Source

Computationally simulated based on parameters derived from Haj Sghaier et al. (2022) [doi:10.3390/plants11212819 ]. https://www.mdpi.com/2223-7747/11/21/2819

srrstats compliance

.

References

Haj Sghaier, A., Tarnawa, Á., Khaeim, H., Kovács, G. P., Gyuricza, C., & Kende, Z. (2022). The Effects of Temperature and Water on the Seed Germination and Seedling Development of Rapeseed (Brassica napus L.). *Plants*,11(21), 2819. https://doi.org/10.3390/plants11212819